Disable Makefile regeneration for openocd.
Seems quite strange, so a detailed explanation why I did this.

OpenOCD is an automake based project, so it includes rules in the final Makefile
to rebuild itself if one of the input files for automake or configure changed.
Just building the release archive of openOCD should never trigger a rebuild of
the Makefile as none of the inputs changed. However building on GitHub failed in
a random manner as sometimes the "make" tool thought a rebuild was necessary and
no autoconfig in a recent enough version was found. This never ever happened on
my local machine.

A closer look revealed that make was really correct. The file
/src/target/Makefile.am was a few milliseconds younger than /Makefile.in right
after depacking the archive. I guess this depends on the order in which the files
were depacked from the source archive.

The first solution coming to mind is simply accepting things as they are and
install automake. Unfortunately openOCD requires automake-1.16, which is not
available in the standard Ubuntu 18.04 repository. So in the end it was easier
to apply a little patch commenting out all the Makefile regeneration stuff.

I still do not know why the problem only occurs on GitHub and not on my local
machine, but after debugging for a few hours I'm happy to have at least an ugly
workaround.